code refactoring: removed logger text view. cleaned code
