+++ Podfile
@@ -1,6 +1,6 @@
 use_frameworks!
 target "nRF Toolbox" do
-    pod 'iOSDFULibrary'#, '~> 4.5'
+    pod 'iOSDFULibrary', '~> 4.6'
     pod 'EVReflection', '~> 5.10'
     pod 'SWRevealViewController', '~> 2.3'
     pod 'CorePlot', '~> 2.2'
